workign crud routes

+++ migrations/20170721184314_users.js
@@ -6,7 +6,7 @@ exports.up = function(knex, Promise) {
     table.string('last_name').notNullable().defaultTo('');
     table.string('email').notNullable().unique();
     table.specificType('hashed_password','char(60)').notNullable();
-    table.integer('phone').notNullable().unique();
+    table.bigInteger('phone').notNullable().unique();
   });
 };
